Move mock test classes to module directory so they can be autoloaded
